Moving all the scanning js
