+++ scanomatic/ui_server_data/js/src/scanning.jsx
